Status processing added. Functions of the downloads(audio and image) are moved to the utils.js
